Integrate Phase 2 and run its tests for the first time
Phase 2 shipped 7,300 lines whose tests had never executed - ruff failed first
in CI, so format, mypy and every test module were skipped rather than failed.
Running them found real defects, not just style:

mypy, first run on this code, five errors. The substantive one is in
score_extraction: `mention` was bound by two consecutive loops, first over
GoldMention and then over EventMention, so every EventMention access in the
second loop was being checked against the wrong type - and .supporting_span,
which GoldMention does not have, type-checked only because nothing ever ran.
Renamed the first binding. Also an unused type: ignore in the cascade, and an
actor-less cluster's [None] fallback typed as list[str].

The ARG002 finding was worth chasing and turned out benign. _buckets took a
score it never used, and _hypothesis took one only to pass it there. The score
was never lost: it lands on CandidateProposal.generator_score. And it must not
reach _buckets, because time_bucket_probabilities is validated as a normalised
distribution - it says when-given-the-event - so scaling it by the candidate's
probability would fail validation outright. Dropped the parameter from both and
said so in the docstring, so the next reader does not re-derive it.

The remaining eight were ordinary lint. Nothing about the Phase 2 design needed
changing; it needed running.

916 tests, coverage 90.0% against the 88% floor, M0 leakage/metamorphic/contract
suites green, make demo end to end.

+++ src/pramaanx/calibration/conformal.py
@@ -105,7 +105,7 @@ def required_positives(alpha: float, delta: float) -> int:
         raise ValueError(f"alpha must be in (0, 1), got {alpha}")
     if not 0.0 < delta < 1.0:
         raise ValueError(f"delta must be in (0, 1), got {delta}")
-    return int(math.ceil(math.log(1.0 / delta) / (2.0 * alpha * alpha)))
+    return math.ceil(math.log(1.0 / delta) / (2.0 * alpha * alpha))
 
 
 class RecallFirstController(BaseRiskController):
